Adding tag decoding for datetimes.
Added tag decoders for both datetime strings and datetime epochs.
Both of these are built on top of the generic data model, and no not
impact the parsing of CBOR data, but rather act as helper functions for
mapping from the generic data model into a more specific semantic value.