Changes built up over several editing sessions
Added a round-tripping test suite. This found several bugs leading to
most of the other changes.

Removed the Bool, Null, and Undefined constructors. These are special
values of the Simple constructor, and are better handled via functions.
This doesn't change what a library can do when building values, but it
is a lot easier to test and verify that everything is working properly.

Smart constructors, these are motivated by the need to make boolean,
null, and undefined values without having more constructors in the
data type messing up the Eq and Ord instances.

Manual Eq instance. This is needed to handle floating values, as those
are represented in the smallest value preserving representation. Doubles
may be Halfs over the wire.

Code formatting.

Limiting exports for some modules, and cleaning up the un-used code.

Refactoring the encoding functions to use more combinators. This
massively reduced the size of the module while losing no functionality.

Linting suggestions from various tools to make things more ideomatic.

Changing the Parser to use a new Uncons class, allowing both lazy and
strict bytestrings to be parsed.

Parser errors are now functions, a-la Trifecta's errors.